+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/SearchEngine.kt
@@ -1,12 +1,21 @@
 package com.mapbox.search
 
 import com.mapbox.search.analytics.AnalyticsService
+import com.mapbox.search.base.core.CoreEngineOptions
+import com.mapbox.search.base.core.CoreLocationProvider
+import com.mapbox.search.base.core.CoreSearchEngine
+import com.mapbox.search.base.core.CoreSearchEngineInterface
+import com.mapbox.search.base.location.LocationEngineAdapter
+import com.mapbox.search.base.location.WrapperLocationProvider
+import com.mapbox.search.base.utils.UserAgentProvider
+import com.mapbox.search.base.utils.extension.mapToCore
 import com.mapbox.search.common.AsyncOperationTask
 import com.mapbox.search.common.concurrent.SearchSdkMainThreadWorker
 import com.mapbox.search.record.IndexableDataProvider
 import com.mapbox.search.record.IndexableRecord
 import com.mapbox.search.result.SearchResult
 import com.mapbox.search.result.SearchSuggestion
+import com.mapbox.search.utils.CompoundCompletionCallback
 import java.util.concurrent.Executor
 
 /**
@@ -20,7 +29,7 @@ import java.util.concurrent.Executor
  *
  * ## Instantiating
  *
- * Call [MapboxSearchSdk.createSearchEngine] to get instance of a [SearchEngine].
+ * Call [SearchEngine.createSearchEngine] to get instance of a [SearchEngine].
  *
  * ## Forward geocoding algorithm
  *
@@ -51,7 +60,7 @@ import java.util.concurrent.Executor
  * If invalid parameters are provided, [RuntimeException] is thrown immediately.
  * Any other error will be propagated to [SearchSuggestionsCallback.onError].
  *
- * @see MapboxSearchSdk.createSearchEngine
+ * @see SearchEngine.createSearchEngine
  */
 public interface SearchEngine {
 
@@ -327,4 +336,93 @@ public interface SearchEngine {
         executor = SearchSdkMainThreadWorker.mainExecutor,
         callback = callback,
     )
+
+    /**
+     * @suppress
+     */
+    public companion object {
+
+        /**
+         * Creates a new instance of the [SearchEngine] with a default [ApiType].
+         * A new instance doesn't have any [IndexableDataProvider] registered by default.
+         *
+         * @param settings [SearchEngine] settings.
+         *
+         * @return a new instance instance of [SearchEngine].
+         * @see createSearchEngineWithBuiltInDataProviders
+         */
+        @JvmStatic
+        public fun createSearchEngine(settings: SearchEngineSettings): SearchEngine {
+            return createSearchEngine(ApiType.GEOCODING, settings)
+        }
+
+        /**
+         * Experimental API, can be changed or removed in the next SDK releases.
+         *
+         * Creates a new instance of the [SearchEngine].
+         * A new instance doesn't have any [IndexableDataProvider] registered by default.
+         *
+         * @param settings [SearchEngine] settings.
+         * @param apiType The type of the API used by the Search Engines.
+         * Note that [ApiType.GEOCODING] is the only available publicly.
+         * You might need to [contact sales](https://www.mapbox.com/contact/sales/) to enable access for other API types.
+         *
+         * @return a new instance instance of [SearchEngine].
+         * @see createSearchEngineWithBuiltInDataProviders
+         */
+        @JvmStatic
+        public fun createSearchEngine(apiType: ApiType, settings: SearchEngineSettings): SearchEngine {
+            return SearchEngineFactory().createSearchEngine(apiType, settings)
+        }
+
+        /**
+         * Creates a new instance of the [SearchEngine] with a default [ApiType] and default data providers (
+         * [com.mapbox.search.record.HistoryDataProvider] and [com.mapbox.search.record.FavoritesDataProvider])
+         * registered by default.
+         *
+         * @param settings [SearchEngine] settings.
+         * @param executor Executor used for events dispatching. By default events are dispatched on the main thread.
+         * @param callback Callback to handle result.
+         *
+         * @return a new instance of [SearchEngine].
+         * @see createSearchEngine
+         */
+        @JvmOverloads
+        @JvmStatic
+        public fun createSearchEngineWithBuiltInDataProviders(
+            settings: SearchEngineSettings,
+            executor: Executor = SearchSdkMainThreadWorker.mainExecutor,
+            callback: CompletionCallback<Unit> = StubCompletionCallback()
+        ): SearchEngine {
+            return createSearchEngineWithBuiltInDataProviders(ApiType.GEOCODING, settings, executor, callback)
+        }
+
+        /**
+         * Experimental API, can be changed or removed in the next SDK releases.
+         *
+         * Creates a new instance of the [SearchEngine] with default data providers (
+         * [com.mapbox.search.record.HistoryDataProvider] and [com.mapbox.search.record.FavoritesDataProvider])
+         * registered by default.
+         *
+         * @param settings [SearchEngine] settings.
+         * @param apiType The type of the API used by the Search Engines. By default [ApiType.GEOCODING] will be used.
+         * Note that [ApiType.GEOCODING] is the only available publicly.
+         * You might need to [contact sales](https://www.mapbox.com/contact/sales/) to enable access for other API types.
+         * @param executor Executor used for events dispatching. By default events are dispatched on the main thread.
+         * @param callback Callback to handle result.
+         *
+         * @return a new instance of [SearchEngine].
+         * @see createSearchEngine
+         */
+        @JvmOverloads
+        @JvmStatic
+        public fun createSearchEngineWithBuiltInDataProviders(
+            apiType: ApiType,
+            settings: SearchEngineSettings,
+            executor: Executor = SearchSdkMainThreadWorker.mainExecutor,
+            callback: CompletionCallback<Unit> = StubCompletionCallback()
+        ): SearchEngine {
+            return SearchEngineFactory().createSearchEngineWithBuiltInDataProviders(apiType, settings, executor, callback)
+        }
+    }
 }