+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/result/IndexableRecordSearchResultImpl.kt
@@ -37,7 +37,7 @@ internal data class IndexableRecordSearchResultImpl(
     override val types: List<SearchResultType>
         get() = listOf(record.type)
 
-    override val categories: List<String>
+    override val categories: List<String>?
         get() = record.categories ?: super.categories
 
     override val makiIcon: String?